+++ CHANGELOG.md
@@ -51,6 +51,7 @@ v1.8.0 [unreleased]
 - [#334](https://github.com/latera/homs/pull/334) Add my/unassigned tasks switch component.
 - [#333](https://github.com/latera/homs/pull/333) Optimize Camunda process definitions fetching.
 - [#332](https://github.com/latera/homs/pull/332) Add table for claiming.
+- [#337](https://github.com/latera/homs/pull/337) Add claim button for widget page.
 
 v1.7.8 [2019-10-03]
 -------------------